Move messages schemas/constants to lib/messages
