+++ apps/web/app/(ee)/api/partner-profile/messages/count/route.ts
@@ -1,5 +1,5 @@
 import { withPartnerProfile } from "@/lib/auth/partner";
-import { countMessagesQuerySchema } from "@/lib/zod/schemas/messages";
+import { countMessagesQuerySchema } from "@/lib/messages/schemas";
 import { prisma } from "@dub/prisma";
 import { NETWORK_PROGRAM_ID } from "@dub/utils";
 import { NextResponse } from "next/server";
