+++ apps/web/app/(ee)/partners.dub.co/(dashboard)/messages/[programSlug]/page-client.tsx
@@ -4,6 +4,7 @@ import { parseActionError } from "@/lib/actions/parse-action-errors";
 import { markProgramMessagesReadAction } from "@/lib/actions/partners/mark-program-messages-read";
 import { messageProgramAction } from "@/lib/actions/partners/message-program";
 import { uploadPartnerMessageAttachmentAction } from "@/lib/actions/partners/upload-partner-message-attachment";
+import { PARTNER_ALLOWED_ATTACHMENT_TYPES } from "@/lib/messages/constants";
 import { constructPartnerLink } from "@/lib/partners/construct-partner-link";
 import { mutatePrefix } from "@/lib/swr/mutate";
 import usePartnerAnalytics from "@/lib/swr/use-partner-analytics";
@@ -12,7 +13,6 @@ import useProgramEnrollment from "@/lib/swr/use-program-enrollment";
 import { useProgramMessages } from "@/lib/swr/use-program-messages";
 import useUser from "@/lib/swr/use-user";
 import { ProgramEnrollmentProps } from "@/lib/types";
-import { PARTNER_ALLOWED_ATTACHMENT_TYPES } from "@/lib/zod/schemas/messages";
 import { INACTIVE_ENROLLMENT_STATUSES } from "@/lib/zod/schemas/partners";
 import { useMessagesContext } from "@/ui/messages/messages-context";
 import { MessagesPanel } from "@/ui/messages/messages-panel";
