+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/messages/[partnerId]/page-client.tsx
@@ -4,13 +4,13 @@ import { parseActionError } from "@/lib/actions/parse-action-errors";
 import { markPartnerMessagesReadAction } from "@/lib/actions/partners/mark-partner-messages-read";
 import { messagePartnerAction } from "@/lib/actions/partners/message-partner";
 import { uploadMessageAttachmentAction } from "@/lib/actions/partners/upload-message-attachment";
+import { PROGRAM_ALLOWED_ATTACHMENT_TYPES } from "@/lib/messages/constants";
 import { mutatePrefix } from "@/lib/swr/mutate";
 import usePartner from "@/lib/swr/use-partner";
 import { usePartnerMessages } from "@/lib/swr/use-partner-messages";
 import useProgram from "@/lib/swr/use-program";
 import useUser from "@/lib/swr/use-user";
 import useWorkspace from "@/lib/swr/use-workspace";
-import { PROGRAM_ALLOWED_ATTACHMENT_TYPES } from "@/lib/zod/schemas/messages";
 import { useMessagesContext } from "@/ui/messages/messages-context";
 import { MessagesPanel } from "@/ui/messages/messages-panel";
 import { ToggleSidePanelButton } from "@/ui/messages/toggle-side-panel-button";
