+++ apps/web/lib/messages/schemas.ts
@@ -1,50 +1,13 @@
+import {
+  MAX_ATTACHMENT_SIZE_BYTES,
+  MAX_ATTACHMENTS_PER_MESSAGE,
+  MAX_MESSAGE_LENGTH,
+} from "@/lib/messages/constants";
 import { MessageType } from "@dub/prisma/client";
 import * as z from "zod/v4";
-import { PartnerSchema } from "./partners";
-import { ProgramSchema } from "./programs";
-import { UserSchema } from "./users";
-
-export const MAX_MESSAGE_LENGTH = 2000;
-
-export const MAX_ATTACHMENTS_PER_MESSAGE = 5;
-
-export const MAX_ATTACHMENT_SIZE_BYTES = 10 * 1024 * 1024; // 10MB
-
-export const MAX_ATTACHMENT_NAME_LENGTH = 255;
-
-export const PROGRAM_ALLOWED_ATTACHMENT_TYPES = [
-  "image/png",
-  "image/jpeg",
-  "image/webp",
-  "application/pdf",
-] as const;
-
-export const PARTNER_ALLOWED_ATTACHMENT_TYPES = [
-  "image/png",
-  "image/jpeg",
-  "image/webp",
-] as const;
-
-export const ATTACHMENT_MIME_TYPE_LABELS: Record<string, string> = {
-  "application/pdf": "PDF",
-  "image/png": "PNG",
-  "image/jpeg": "JPG",
-  "image/webp": "WEBP",
-};
-
-// MIME types that are safe to render as inline image previews.
-export const PREVIEWABLE_IMAGE_TYPES = new Set([
-  "image/png",
-  "image/jpeg",
-  "image/webp",
-]);
-
-export const ATTACHMENT_MIME_TYPE_COLOR: Record<string, string> = {
-  "application/pdf": "bg-red-600",
-  "image/png": "bg-blue-600",
-  "image/jpeg": "bg-blue-500",
-  "image/webp": "bg-blue-500",
-};
+import { PartnerSchema } from "../zod/schemas/partners";
+import { ProgramSchema } from "../zod/schemas/programs";
+import { UserSchema } from "../zod/schemas/users";
 
 const messageTextSchema = z.string().max(MAX_MESSAGE_LENGTH);
