+++ apps/web/ui/shared/message-input.tsx
@@ -1,13 +1,13 @@
 import {
-  getAttachmentTypeLabel,
-  isPreviewableImageType,
-} from "@/lib/messages/utils";
-import {
   ATTACHMENT_MIME_TYPE_COLOR,
   MAX_ATTACHMENTS_PER_MESSAGE,
   MAX_MESSAGE_LENGTH,
-  messageAttachmentInputSchema,
-} from "@/lib/zod/schemas/messages";
+} from "@/lib/messages/constants";
+import { messageAttachmentInputSchema } from "@/lib/messages/schemas";
+import {
+  getAttachmentTypeLabel,
+  isPreviewableImageType,
+} from "@/lib/messages/utils";
 import {
   ArrowTurnLeft,
   Button,
